+++ src/screens/modules/schedule/AddEventScreen.js
@@ -13,8 +13,8 @@ import {
     AppFormBox,
     KeyboardView,
     SubmitButton
-} from '../components';
-import { addEvent } from '../utils/api_handler/events';
+} from '../../../components';
+import { addEvent } from '../../../utils/api_handler/events';
 
 const validationSchema = Yup.object().shape({
     eventTitle: Yup.string()
